requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
fastapi
uvicorn
sqlalchemy
pyodbc
python-dotenv
pydantic
pandas
python-multipart
azure-storage-file-datalake
deltalake
requests
duckdb
pydantic_settings
deltalake[pyarrow]
aiohttp
